refactor: 리뷰 컨트롤러 통합
